+++ control-plane/api/v1alpha1/mesh_webhook_test.go
@@ -63,6 +63,23 @@ func TestValidateMesh(t *testing.T) {
 			expAllow:      false,
 			expErrMessage: "mesh resource name must be \"mesh\"",
 		},
+		"validation rejects": {
+			existingResources: nil,
+			newResource: &Mesh{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: common.Mesh,
+				},
+				Spec: MeshSpec{
+					TLS: &MeshTLSConfig{
+						Incoming: &MeshDirectionalTLSConfig{
+							TLSMinVersion: "foo",
+						},
+					},
+				},
+			},
+			expAllow:      false,
+			expErrMessage: "mesh.consul.hashicorp.com \"mesh\" is invalid: spec.tls.incoming.tlsMinVersion: Invalid value: \"foo\": must be one of \"TLS_AUTO\", \"TLSv1_0\", \"TLSv1_1\", \"TLSv1_2\", \"TLSv1_3\", \"\"",
+		},
 	}
 	for name, c := range cases {
 		t.Run(name, func(t *testing.T) {